+++ src/main/java/com/creddit/credditmainserver/dto/request/PostRequestDto.java
@@ -2,7 +2,6 @@ package com.creddit.credditmainserver.dto.request;
 
 import com.creddit.credditmainserver.domain.Member;
 import com.creddit.credditmainserver.domain.Post;
-import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,22 +9,30 @@ import lombok.NoArgsConstructor;
 import javax.validation.constraints.NotBlank;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostUpdateRequestDto {
+@NoArgsConstructor
+public class PostRequestDto {
 
+    @NotBlank(message = "제목은 빈 칸일 수 없습니다.")
     private String title;
+
+    @NotBlank(message = "내용은 빈 칸일 수 없습니다.")
     private String content;
+
     private String imgName;
 
     @Builder
-    public PostUpdateRequestDto(String title, String content, String imgName) {
+    public PostRequestDto(String title, String content) {
         this.title = title;
         this.content = content;
+    }
+
+    public void addImgName(String imgName){
         this.imgName = imgName;
     }
 
-    public Post toEntity(){
+    public Post toEntity(Member member){
         return Post.builder()
+                .member(member)
                 .title(title)
                 .content(content)
                 .imgName(imgName)
